requirements.txt
aiohttp==3.7.4.post0
altgraph==0.17
anaconda==0.0.1.1
appdirs==1.4.4
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
auto-py-to-exe==2.8.0
backcall==0.2.0
beautifulsoup4==4.9.3
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
comtypes==1.1.8
cycler==0.10.0
DateTime==4.3
decorator==5.0.7
discord.py==1.7.1
distlib==0.3.1
docopt==0.6.2
Eel==0.12.4
emoji==0.6.0
filelock==3.0.12
future==0.18.2
geographiclib==1.50
geopy==2.1.0
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.0.0
h5py==3.1.0
idna==2.10
ipython-genutils==0.2.0
isort==5.7.0
jedi==0.18.0
Js2Py==0.70
Keras==2.4.3
keyboard==0.13.5
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.3.3
matplotlib-inline==0.1.2
mccabe==0.6.1
MouseInfo==0.1.3
multidict==5.1.0
numpy==1.19.3
opencv-python==4.4.0.46
packaging==20.8
pandas==1.2.0
parso==0.8.2
pefile==2019.4.18
pickleshare==0.7.5
Pillow==8.0.1
playsound==1.2.2
plyer==2.0.0
prompt-toolkit==3.0.18
PyAutoGUI==0.9.52
pycparser==2.20
pyemojify==0.2.0
pyFirmata==1.1.0
pygame==2.0.1
PyGetWindow==0.0.9
pyglet==1.5.13
Pygments==2.8.1
pyinstaller==4.2
pyjokes==0.6.0
pyjsparser==2.7.1
pylint==2.6.2
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.1
PyPrind==2.11.2
PyRect==0.1.4
PyScreeze==0.1.26
pyserial==3.5
pySmartDL==1.3.4
python-dateutil==2.8.1
pyttsx3==2.90
PyTweening==1.0.3
pytz==2020.5
pywhatkit==3.8
PyYAML==5.3.1
requests==2.25.1
rotate-screen==0.1.0
scipy==1.5.4
six==1.15.0
soupsieve==2.1
toml==0.10.2
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.2
virtualenv==20.3.1
wcwidth==0.2.5
weather-forecast==0.49
whichcraft==0.6.1
wikipedia==1.4.0
wrapt==1.12.1
yarl==1.6.3
zope.event==4.5.0
zope.interface==5.2.0
